Fix DST Bug
Fixed DST not working bug by init-ing the tm_isdst variable to "-1". Then we can call mktime on the struct to get the OS to set the value of DST based on the system settings. Removed the isdst bypass and added messages to state if DST is active or not.